Remove entire code base of 'different-react-versions-16-18-typescript' project
All the files related to the 'different-react-versions-16-18-typescript' project have been deleted. This includes components, configurations, scripts, and documentations. This major cleanup suggests a decision to retire or significantly revamp this project.
