+++ different-react-versions-16-17-typescript/app2/package.json
@@ -1,9 +1,12 @@
 {
   "name": "different-react-versions-16-17-typescript_app2",
+  "description": "React 17",
   "version": "0.0.0",
   "scripts": {
-    "start": "webpack-cli serve",
-    "build": "webpack --mode production",
+    "start": "rspack serve",
+    "legacy:start": "webpack-cli serve",
+    "build": "rspack --mode production",
+    "legacy:build": "webpack --mode production",
     "serve": "serve dist -p 3002",
     "clean": "rm -rf dist"
   },
@@ -24,6 +27,9 @@
     "typescript": "4.8.4",
     "webpack": "5.89.0",
     "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.15.1"
+    "webpack-dev-server": "4.15.1",
+    "@rspack/core": "^0.4.5",
+    "@rspack/cli": "^0.4.5",
+    "@rspack/dev-server": "^0.4.5"
   }
 }
