+++ different-react-versions-16-17-typescript/app2/webpack.config.js
@@ -52,16 +52,10 @@ const webpackConfig = {
         {
           react: {
             import: 'react', // the "react" package will be used a provided and fallback module
-            shareKey: 'newReact', // under this name the shared module will be placed in the share scope
-            shareScope: 'default', // share scope with this name will be used
+            shareKey: 'react', // under this name the shared module will be placed in the share scope
+            shareScope: 'modern', // share scope with this name will be used
             singleton: true, // only a single version of the shared module is allowed
           },
-          // reactNew: {
-          //   import: "react", // the "react" package will be used a provided and fallback module
-          //   shareKey: "reactNew", // under this name the shared module will be placed in the share scope
-          //   shareScope: "modern", // share scope with this name will be used
-          //   singleton: true, // only a single version of the shared module is allowed
-          // },
         },
       ],
     }),
