+++ different-react-versions-16-17-typescript/package.json
@@ -1,11 +1,16 @@
 {
   "name": "different-react-versions-16-17-typescript",
+  "description": "TypeScript example demos the ability to load two separate versions of react.",
   "version": "0.0.0",
   "scripts": {
     "start": "pnpm --filter different-react-versions-16-17-typescript_* --parallel start",
+    "legacy:start": "pnpm --filter different-react-versions-16-17-typescript_* --parallel legacy:start",
     "build": "pnpm --filter different-react-versions-16-17-typescript_* build",
+    "legacy:build": "pnpm --filter different-react-versions-16-17-typescript_* legacy:build",
     "serve": "pnpm --filter different-react-versions-16-17-typescript_* --parallel serve",
-    "clean": "pnpm --filter different-react-versions-16-17-typescript_* --parallel clean"
+    "clean": "pnpm --filter different-react-versions-16-17-typescript_* --parallel clean",
+    "legacy:e2e:ci": "pnpm run legacy:build",
+    "e2e:ci": "pnpm run build"
   },
   "devDependencies": {}
 }
