+++ different-react-versions-16-17/package.json
@@ -8,7 +8,9 @@
     "legacy:build": "pnpm --filter different-react-versions-16-17_* legacy:build",
     "build": "pnpm --filter different-react-versions-16-17_* build",
     "serve": "pnpm --filter different-react-versions-16-17_* --parallel serve",
-    "clean": "pnpm --filter different-react-versions-16-17_* --parallel clean"
+    "clean": "pnpm --filter different-react-versions-16-17_* --parallel clean",
+    "legacy:e2e:ci": "pnpm run legacy:build",
+    "e2e:ci": "pnpm run build"
   },
   "devDependencies": {}
 }
