--- different-react-versions-16-18-typescript/app1/index.d.ts
@@ -1,7 +1,0 @@
-declare module 'app2/Button' {
-  export * from 'app2/src/components/Button';
-}
-
-declare module 'app2/ModernComponent' {
-  export * from 'app2/src/components/ModernReactComponent';
-}
