--- different-react-versions-16-18-typescript/app1/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@different-react-versions-16-18-typescript/app1",
-  "version": "0.0.0",
-  "scripts": {
-    "start": "webpack-cli serve",
-    "build": "webpack --mode production",
-    "serve": "serve dist -p 3001",
-    "clean": "rm -rf dist"
-  },
-  "dependencies": {
-    "react": "16.14.0",
-    "react-dom": "16.14.0"
-  },
-  "devDependencies": {
-    "@babel/core": "7.23.5",
-    "@babel/plugin-proposal-class-properties": "7.18.6",
-    "@babel/preset-react": "7.23.3",
-    "@babel/preset-typescript": "7.23.3",
-    "@types/react": "16.14.55",
-    "@types/react-dom": "16.9.24",
-    "babel-loader": "8.2.5",
-    "html-webpack-plugin": "5.6.0",
-    "serve": "14.2.1",
-    "typescript": "4.8.4",
-    "webpack": "5.89.0",
-    "webpack-cli": "4.9.2",
-    "webpack-dev-server": "4.15.1"
-  }
-}
