--- different-react-versions-16-18-typescript/app1/public/index.html
@@ -1,8 +1,0 @@
-<html>
-  <head>
-    <script src="<%= htmlWebpackPlugin.options.app2RemoteEntry %>"></script>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
