--- different-react-versions-16-18-typescript/app1/src/components/ComponentWithHook.tsx
@@ -1,11 +1,0 @@
-import React, { useEffect } from 'react';
-
-const ComponentWithHook = () => {
-  React.useEffect(() => {
-    console.log('some effect from app1');
-  }, []);
-
-  return <span>This should break, no hooks supported in this app.</span>;
-};
-
-export default ComponentWithHook;
