--- different-react-versions-16-18-typescript/app2/.babelrc
@@ -1,9 +1,0 @@
-{
-  "plugins": [
-    "@babel/plugin-proposal-class-properties"
-  ],
-  "presets": [
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ]
-}
