--- different-react-versions-16-18-typescript/app2/public/index.html
@@ -1,5 +1,0 @@
-<html>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
