--- different-react-versions-16-18-typescript/app2/src/bootstrap.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './components/App';
-
-ReactDOM.render(<App />, document.getElementById('root'));
