--- different-react-versions-16-18-typescript/app2/src/components/App.tsx
@@ -1,12 +1,0 @@
-import LocalButton from './Button';
-import React from 'react';
-
-const App = () => (
-  <div>
-    <h1>Basic Host-Remote</h1>
-    <h2>App 2</h2>
-    <LocalButton />
-  </div>
-);
-
-export default App;
