--- different-react-versions-16-18-typescript/app2/src/index.ts
@@ -1,1 +1,0 @@
-import('./bootstrap');
