--- different-react-versions-16-18-typescript/app2/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {},
-  "include": ["src/**/*", "**/*.d.ts"],
-  "exclude": ["node_modules"]
-}
