--- different-react-versions-16-18-typescript/lerna.json
@@ -1,5 +1,0 @@
-{
-  "version": "0.0.0",
-  "npmClient": "yarn",
-  "useWorkspaces": true
-}
