--- different-react-versions-16-18-typescript/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "different-react-versions-16-18-typescript",
-  "private": true,
-  "version": "0.0.0",
-  "workspaces": [
-    "app1",
-    "app2"
-  ],
-  "scripts": {
-    "start": "pnpm --filter --parallel start",
-    "build": "pnpm --filter build",
-    "serve": "pnpm --filter --parallel serve",
-    "clean": "pnpm --filter --parallel clean"
-  },
-  "devDependencies": {}
-}
