--- different-react-versions-16-18-typescript/sandbox.config.json
@@ -1,8 +1,0 @@
-{
-  "template": "node",
-  "container": {
-    "port": 3001,
-    "startScript": "start",
-    "node": "12"
-  }
-}
