--- different-react-versions-16-18-typescript/tsconfig.base.json
@@ -1,23 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "esnext",
-    "jsx": "react",
-    "declaration": false,
-    "importHelpers": true,
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "skipDefaultLibCheck": true,
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "baseUrl": ".",
-    "paths": {
-      "@app1/*": ["./app1/src/*"],
-      "@app2/*": ["./app2/src/*"]
-    }
-  },
-  "exclude": ["node_modules", "tmp"]
-}
