+++ different-react-versions-16-18/app2/package.json
@@ -1,5 +1,6 @@
 {
   "name": "different-react-versions-16-18_app2",
+  "description": "React 18",
   "version": "0.0.0",
   "devDependencies": {
     "@babel/core": "7.23.5",
@@ -10,7 +11,10 @@
     "serve": "14.2.1",
     "webpack": "5.89.0",
     "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.15.1"
+    "webpack-dev-server": "4.15.1",
+    "@rspack/core": "^0.4.5",
+    "@rspack/cli": "^0.4.5",
+    "@rspack/dev-server": "^0.4.5"
   },
   "scripts": {
     "start": "webpack-cli serve",
