+++ different-react-versions-16-18/app2/rspack.config.js
@@ -1,0 +1,80 @@
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const { ModuleFederationPlugin } = require('@rspack/core').container;
+const path = require('path');
+
+/**
+ * @type {import('webpack').Configuration}
+ **/
+const webpackConfig = {
+  entry: './src/index',
+  mode: 'development',
+  devServer: {
+    static: {
+      directory: path.join(__dirname, 'dist'),
+    },
+    hot:true,
+    port: 3002,
+  },
+  output: {
+    publicPath: 'auto',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.m?js$/,
+        type: 'javascript/auto',
+        resolve: {
+          fullySpecified: false,
+        },
+      },
+      {
+        test: /\.js$/,
+        include: path.resolve(__dirname, 'src'),
+        use: {
+          loader: 'builtin:swc-loader',
+          options: {
+            jsc: {
+              parser: {
+                syntax: 'ecmascript',
+                jsx: true,
+              },
+              transform: {
+                react: {
+                  runtime: 'automatic',
+                },
+              },
+            },
+          },
+        },
+      },
+    ],
+  },
+  plugins: [
+    new ModuleFederationPlugin({
+      name: 'app2',
+      library: { type: 'var', name: 'app2' },
+      filename: 'remoteEntry.js',
+      exposes: {
+        './Button': './src/components/Button',
+        './ModernComponent': './src/components/ModernReactComponent',
+        './newReact': require.resolve('react'),
+        './newReactDOM': require.resolve('react-dom'),
+      },
+      shared: [
+        'react-dom',
+        {
+          react: {
+            import: 'react', // the "react" package will be used a provided and fallback module
+            shareScope: 'modern', // share scope with this name will be used
+            singleton: true, // only a single version of the shared module is allowed
+          },
+        },
+      ],
+    }),
+    new HtmlWebpackPlugin({
+      template: './public/index.html',
+    }),
+  ],
+};
+
+module.exports = webpackConfig;
